:sparkles: ナビバーに認証処理のボタンを移動
