+++ src/router/index.ts
@@ -1,7 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Home from '@/views/Home.vue'
 import Search from '@/views/Search.vue'
-import Signin from '@/views/Signin.vue'
 // import WishView from '@/views/Wish.vue'
 // import DoneView from '@/views/Done.vue'
 
@@ -12,11 +11,6 @@ const router = createRouter({
       path: '/',
       name: 'home',
       component: Home
-    },
-    {
-      path: '/signin',
-      name: 'signin',
-      component: Signin
     },
     {
       path: '/search',
